Move the encoding to encoder.py
